BlockLoader reorg

+++ src/xp/Calendar.cc
@@ -4,9 +4,12 @@ class Calendar : public XDLType("Calendar") {
 private:
 	std::string name;                // name of this calendar
 	Date start;
-	uint32_t weekendDays : 7;
+	uint32_t weekendDays : 7; // 0b1000001
 	std::vector<uint16_t> yearStart; // Index into the holiday table for each year in the calendar
-	std::vector<uint8_t> holidays;   // offset for each holiday in a given year
+	std::vector<uint16_t> holidays;   // offset for each holiday in a given year
+									  // example: start Jan 1 2020
+									  // yearStart = 0  10
+									  // holidays: 0 45 ... july 4 ....  dec.25,  next year starts 0 ...
 public:
 	Calendar(const char filename[]);
 	Calendar(const std::string& name
